Top menu drawing correctly. Version description move to oc3_project_version.hpp add StringHelper::debug for format debug message and out it in console. house_model.csv convert to json format placed in resources/house.model. house model now load in HouseSpecHelper::load. Json parser now may read json files with comment in /* */ block. oc3_model_loader.cpp/hpp delete from project.
Former-commit-id: 0df46d2f76aab08427c64f0650f00245099cbd95

+++ oc3_listbox.cpp
@@ -573,7 +573,7 @@ void ListBox::beforeDraw( GfxEngine& painter)
          ListBoxItem& refItem = _d->items[i];
 
          if( frameRect.LowerRightCorner.getY() >= 0 &&
-             frameRect.UpperLeftCorner.getY() <= getHeight() )
+             frameRect.UpperLeftCorner.getY() <= (int)getHeight() )
          {
            refItem.setState( _GetCurrentItemState( i, hl ) );
            
@@ -638,7 +638,7 @@ void ListBox::_RecalculateScrollPos()
 	{
         _d->scrollBar->setPos( _d->scrollBar->getPos() + selPos );
 	}
-	else if (selPos > getHeight() - _d->itemHeight)
+	else if (selPos > (int)getHeight() - _d->itemHeight)
 	{
         _d->scrollBar->setPos( _d->scrollBar->getPos() + selPos - getHeight() + _d->itemHeight );
 	}
